+++ src/HasStates.php
@@ -79,8 +79,6 @@ trait HasStates
                     $state->setField($stateConfig->field);
                 }
 
-                $state->setField($stateConfig->field);
-
                 $model->setAttribute(
                     $stateConfig->field,
                     $state
@@ -167,13 +165,9 @@ trait HasStates
         return $this->{$field}->transitionTo($state);
     }
 
-    public function transitionableStates(string $fromClass, ?string $field = null): array
+    public function transitionableStates(string $fromClass, string $field): array
     {
         $stateConfig = self::getStateConfig();
-
-        if ($field === null && count($stateConfig) > 1) {
-            throw InvalidConfig::fieldNotFound($fromClass, $this);
-        }
 
         $field = $field ?? reset($stateConfig)->field;
 
@@ -215,7 +209,7 @@ trait HasStates
     /**
      * @return \Spatie\ModelStates\StateConfig[]
      */
-    private static function getStateConfig(): array
+    public static function getStateConfig(): array
     {
         if (static::$stateFields === null) {
             static::$stateFields = [];